+++ CHANGELOG.md
@@ -9,7 +9,9 @@
 - [#1050](https://github.com/okta/okta-auth-js/pull/1050) Removes `userAgent` field from oktaAuth instance
 - [#1014](https://github.com/okta/okta-auth-js/pull/1014) Shared transaction storage is automatically cleared on success and error states. Storage is not cleared for "terminal" state which is neither success nor error.
 - [#1051](https://github.com/okta/okta-auth-js/pull/1051) Removes `useMultipleCookies` from CookieStorage options
-- [#1059](https://github.com/okta/okta-auth-js/pull/1059) Sets signOut option `clearTokensAfterRedirect` default value to true
+- [#1059](https://github.com/okta/okta-auth-js/pull/1059) 
+  - Removes signOut option `clearTokensAfterRedirect`
+  - Adds signOut option `clearTokensAfterRedirect` (default: `false`) to remove local tokens before logout redirect happen
 
 ### Features